Moving core-command to ./cmd
