+++ cmd/core-command/res/configuration-docker.json
