+++ cmd/core-command/res/configuration.json
